backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
Flask-Migrate==4.1.0
PyMySQL==1.1.1
python-dotenv==1.1.0
waitress==3.0.2
gunicorn==23.0.0; sys_platform != 'win32'
